test: dedupe RouterClient to satisfy SonarCloud
